Split apart world generation module
